+++ packages/eyes-sdk-core/test/unit/server/SessionStartInfo.spec.js
@@ -17,10 +17,10 @@ const {
   AccessibilityVersion,
   AccessibilityRegionType,
   AccessibilityRegionByRectangle,
-  EyesBase,
   CheckTarget,
   MatchWindowTask,
 } = require('../../../index')
+const {EyesBaseImpl} = require('../../testUtils')
 
 describe('SessionStartInfo', () => {
   it('TestSerialization', () => {
@@ -109,7 +109,7 @@ describe('SessionStartInfo', () => {
         .enablePatterns(enablePatterns)
         .ignoreDisplacements(ignoreDisplacements)
 
-      const eyes = new EyesBase()
+      const eyes = new EyesBaseImpl()
       const task = new MatchWindowTask(true, true, true, true, eyes, true)
       const imageMatchSettings = await task.createImageMatchSettings(settings, null)
 
@@ -130,7 +130,7 @@ describe('SessionStartInfo', () => {
         .useDom(useDom)
         .enablePatterns(enablePatterns)
 
-      const eyes = new EyesBase()
+      const eyes = new EyesBaseImpl()
       const configuration = eyes.getConfiguration()
       configuration.setIgnoreDisplacements(ignoreDisplacements)
       eyes.setConfiguration(configuration)
@@ -152,7 +152,7 @@ describe('SessionStartInfo', () => {
     it(`TestConfigurationSerialization (${useDom}, ${enablePatterns}, ${ignoreDisplacements})`, async () => {
       const settings = CheckTarget.window().fully()
 
-      const eyes = new EyesBase()
+      const eyes = new EyesBaseImpl()
       const configuration = eyes.getConfiguration()
       configuration.setUseDom(useDom)
       configuration.setEnablePatterns(enablePatterns)