+++ packages/eyes-selenium-3/src/ScrollPositionProvider.js
@@ -1,8 +1,8 @@
 ;(function() {
   'use strict'
 
-  var ArgumentGuard = require('eyes.utils').ArgumentGuard
-  var PositionProvider = require('eyes.sdk').PositionProvider
+  var ArgumentGuard = require('@applitools/eyes-common-legacy').ArgumentGuard
+  var PositionProvider = require('@applitools/eyes-sdk-core-legacy').PositionProvider
   var EyesSeleniumUtils = require('./EyesSeleniumUtils').EyesSeleniumUtils
 
   /**