+++ packages/eyes-selenium-3/src/capture/SafariScreenshotImageProvider.js
@@ -1,8 +1,8 @@
 'use strict'
 
 const {By} = require('selenium-webdriver')
-const {OSNames} = require('eyes.utils')
-const {ImageProvider, MutableImage} = require('eyes.sdk')
+const {OSNames} = require('@applitools/eyes-common-legacy')
+const {ImageProvider, MutableImage} = require('@applitools/eyes-sdk-core-legacy')
 
 const {ScrollPositionProvider} = require('../ScrollPositionProvider')