+++ packages/visual-grid-client/test/util/FakeEyesWrapper.js
@@ -388,11 +388,11 @@ class FakeEyesWrapper extends EventEmitter {
   }
 
   setBaseAgentId(value) {
-    this.agentId = value
+    this.baseAgentId = value
   }
 
   getBaseAgentId() {
-    return this.agentId || 'fake wrapper'
+    return this.baseAgentId || 'fake wrapper'
   }
 
   getApiKey() {