+++ packages/dom-utils/CHANGELOG.md
@@ -3,6 +3,15 @@
 ## Unreleased
 
 
+## 4.7.12 - 2020/4/23
+
+- updated to @applitools/eyes-common@3.21.1
+
+## 4.7.11 - 2020/4/22
+
+- updated to @applitools/dom-capture@7.1.5
+- updated to @applitools/eyes-common@3.21.0
+
 ## 4.7.10
 
 - updated eyes-common