+++ packages/eyes-sdk-core-legacy/CHANGELOG.md
@@ -4,6 +4,11 @@
 
 - support accessibility validation
 
+## 1.0.0 - 20/4/2020
+
+- new package name: @applitools/eyes-sdk-core-legacy
+- added SDK agent id header for eyes server requests
+
 ## 5.0.0 - 29/3/2020
 
 - RunningSession is no longer determined to be new according to startSession's response status, but rather by the response's payload of `isNew`. ([Trello](https://trello.com/c/60Rm4xXG/240-support-future-long-running-tasks))