+++ packages/eyes-sdk-core/CHANGELOG.md
@@ -5,6 +5,14 @@
 - add AccessibilityVersion enum
 - remove accessibilityLevel from checkSettings
 
+## 9.1.2 - 2020/4/23
+
+- updated to @applitools/eyes-common@3.21.1
+
+## 9.1.1 - 2020/4/19
+
+- added SDK agent id header for eyes server requests
+
 ## 9.1.0 - 2020/4/5
 
 - updated to @applitools/eyes-common@3.21.0